Fix javac path escaping for wildcard in cmake.
When CMAKE_CURRENT_BINARY_DIR contains spaces, the quotes are placed around the "*.java" wildcard passed as the input files to javac. Some versions of javac cannot deal with this and can only deal with * when not in quotes.

This change forces the quotes to only be around the directory section of the path and not the wildcard.